requirements.txt
annotated-types==0.7.0
anyio==4.4.0
api.video==1.3.2
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
moviepy==1.0.3
mysql-connector-python==9.0.0
nodeenv==1.9.1
numpy==2.0.0
opencv-python==4.10.0.84
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
prisma==0.13.1
proglog==0.1.10
PyAutoGUI==0.9.54
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python3-xlib==0.15
pytweening==1.2.0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
selenium==4.22.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
StrEnum==0.4.15
tomlkit==0.13.0
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
validators==0.32.0
watchfiles==0.22.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
